rename: Rename from components to base
