--- android/app/src/main/java/com/littlekakao/travelblahblah/ui/base/BaseActivity.kt
@@ -1,4 +1,0 @@
-package com.littlekakao.travelblahblah.ui.base
-
-abstract class BaseActivity {
-}
