+++ android/app/src/main/java/com/littlekakao/travelblahblah/ui/base/BaseText.kt
@@ -1,11 +1,11 @@
-package com.dbinc.uface.ui.components.composable
+package com.littlekakao.travelblahblah.ui.base
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.text.font.FontWeight
 
 @Composable
-fun TitleText(head:String) {
+fun BoldText(head:String) {
     Text(
         text = head,
         fontWeight = FontWeight.Bold,
@@ -13,7 +13,7 @@ fun TitleText(head:String) {
 }
 
 @Composable
-fun SubtitleText(item:String) {
+fun SemiBoldText(item:String) {
     Text(
         text = item,
         fontWeight = FontWeight.SemiBold
